Experimenting/GCE/requirement.txt
google-cloud-logging
gunicorn
